net: bus: add the AF_BUS socket address family
AF_BUS is a new socket address family that allows both unicast and multicast IPC
on a local machine with total ordering for messages (every process on the same
bus sees each message in the same order).

A process can create buses to which other processes can connect and communicate
with each other by sending messages. Processes' addresses are automatically
assigned by the bus on connect and are unique. Messages can be sent either to a
process' unique address or to a bus multicast addresses.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>