net: bus: Force to connect to the bus master socket on connect()
Only connect to the bus master is allowed. The client doesn't need to set the
destination address on connect, it only has to fill the family and sbus_path
fields and omit the client address. So the kernel has to fill this field with
the bus master socket destination address.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>