net: bus: Fix memory corruption bug due bad memory allocation
The sockaddr_bus structure is not word sized aligned and the compiler can add
padding data to optimize memory access.

To save memory, AF_BUS used to allocate only the required amount of memory to
copy the address information instead of allocating a full sockaddr_bus structure
but this had the unfortunate effect that padding information was copied instead
of the real data, so now the sizeof(struct sockaddr_bus) is allocated and copied
even though less memory is needed, to be sure that all the information is
correctly copied.

This patch fixes a bug due bus_bind() allocating addr_len instead of
sizeof(struct sockaddr_bus). If addr_len was less than sockaddr_bus size, then
the sockaddr_bus information was copied on non-reserved memory corrupting it.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>