net: bus: send packets whose destination is unknown to the bus master by default
When a sender specifies a destination which is unknown, AF_BUS should use as a
default destination the bus master. So, the packet is sent to the socket peer
(the accepted socket descriptor).

If the bus master sends a packet to an unknown destination using one of the
accepted socket descriptors, then the packet is resent to the same socket
descriptor that the bus master used to send the message.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>