net: bus: Fix hash bigger than BUS_HASH_SIZE bug
For some AF_BUS addresses the calculated hash was bigger than BUS_HASH_SIZE.
This was due calculating the hash with bus_fold_hash() and XOR'ing with the
AF_BUS address to make the hash more disperse. But bus_fold_hash() returns a
BUS_HASH_SIZE modulo and applying the bitwise operation with the AF_BUS address
can give us a value bigger than BUS_HASH_SIZE.

So, let's just not use the AF_BUS address to calculate the hash.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>