net: bus: Fix bug that non-joined peers could receive unicast messages
only AF_BUS peers that were joined to the bus can send and receive unicast
messages from other peers. non-authenticated clients can only communicate with
the bus master. This was checked for the sender but not for the destination
socket, which means that non-joined peers could receive unicast messages sent
by joined peers.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>