net: bus: decrement the destionation socket refcount on bus_dgram_sendmsg()
bus_find_socket_byaddress() called by bus_find_other() grabs the socket refcount
before returning.

If the destination socket is not authenticated, it cannot receive the packet and
the sender peer socket is used should be used instead.

But before doing this, the non-authenticated destination socket reference count
has to be decremented. Otherwise this leads to a non-balanced refcount in the
kernel avoiding to remove the bus kernel module:

ERROR: Module bus is in use

Solve the issue but decreasing the refcount on the destination socket.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>