net: bus: Check if dgram_send_dgram_finish() was called
The function dgram_send_dgram_finish() function is only called if all
the netfilter hooks returned NF_ACCEPT. If one of the hook functions
returned NF_DROP to discard the packet, then dgram_send_dgram_finish()
is not called and NF_HOOK() returns -EPERM.

Some cleanup is made on dgram_send_dgram_finish(), so it has to be done
on dgram_send_dgram() if that function was not called.

On drop, nf_hook_slow() frees the sk_buff so that step is not necessary.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>