net: bus: add SIOCSINQ ioctl() to change socks maximum send buffer
A sockets transmit queued bytes are limited by a maximum send buffer size.

This parameter (sysctl_wmem_max) can be modified at runtime using the sysctl
interface on /proc/sys/net/core/wmem_default.

The problem is that this is a global paramter an AF_BUS needs a more fine
grained control over the maximum send buffer for each bus.

This patch adds a new ioctl() command for AF_BUS sockets that allows the
bus master socket to modify this parameter for the sockets connected to it.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>