net: bus: don't check for socket type on bus_inq_len()
SIOCINQ ioctl() command returns the amount of queued unread data in
the receive buffer.

It checked the socket type and returned the len of the peek packet in
case of SOCK_DGRAM or it walked over the  skb receive_queue list and
returned the sum of each packet len in case of SOCK_[STREAM | SEQPACKET].

AF_BUS only supports SOCK_SEQPACKET sock type so this check is not needed.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>