net: bus: sanity check user-space provided sockaddr_bus before using it
To send a packet to a specific destination, user-space provide a struct
sockaddr_bus representing the AF_BUS address of the destination.

Also, it has to provide the length of the address. Both are provided on
the message header struct msghdr *msg.

If only the address is provided but no length, then the address is not
valid and is set to NULL.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>