net: bus: don't allow peers to join the bus twice
AF_BUS bus master can allow clients to connect to a bus by means of the
BUS_JOIN_BUS socket option. To enforce that is the bus master and not
the clients who are allowed to do the call, only accepted socket
descriptors can be used.

But a check whether a current client had already joined the bus was not
made before adding to the bus, so let's check if is already authenticated
and inform user-space with an error if tries to join an already joined peer.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>