net: bus: ensure total order by making multicast delivery atomic
AF_BUS requires that packet total ordering. To meet this requirement the
multicast sending of packets has to be atomic. Also, packets can't be
dropped due a multicast recipient having its receive queue full.

If this happen, we put the sender to sleep and to wait on the recipient
wait queue.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>