net: bus: fix unbalanced socket refefence counting bug
sock_hold() and sock_put() are used to grab and release a socket
reference counting.

When the last reference to the socket is released, then the socket can
be freed. But an unbalanced socket refcounting can try to release a sock
while it is still alive. This lead to the following bug:

[  241.228957] Attempt to release alive bus socket: ffff88006d679380

The problem was that we only release the socket reference on unicast
since multicast has to release the socket later when doing the actual
packet delivery.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>