net: bus: fix NULL reference on bus_destruct_scm() bug
The function bus_scm_to_skb() attach ancillary data to a skb
(credentials, file descriptors, etc). It's also set the skb destructor
function to bus_destruct_scm() to cleanup this data.

Unfortunately this function assumed that the socket buffer also has an
associated socket. This is true on unix sockets but on AF_BUS we attach
the ancillary data before associating a socket with the skb. This is
necessary since bus_scm_to_skb() can sleep trying to allocate memory and
the bus peer list to obtain the destination socks is traversed atomically.

So, skb->sk can be NULL in bus_scm_to_skb() and this condition should be
checked when trying to access a sk_buff socket.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>